flask_api_utils/api_requirements.txt
flask
pandasai>=3.0.0b2
flask_cors
python-decouple 
mysql
mysql-connector-python